add documentation for ChaosGizmo
