+++ docs/src/api.md
@@ -13,3 +13,9 @@ All APIs of LiftAndLearn listed in a unstructured manner.
 Modules = [LiftAndLearn]
 Order = [:module, :function, :macro]
 ```
+
+# Submodules
+
+```@docs
+LiftAndLearn.ChaosGizmo
+```
