+++ docs/src/manual/ChaosGizmo/chaosgizmo.md
@@ -1,0 +1,6 @@
+# ChaosGizmo
+
+This is a submodule inside the Lift and Learn framework targeted to analyze chaotic reduced order models. It includes the analysis of the following concepts
+
+- Lyapunov exponents
+- Kaplan-Yorke dimensions
