+++ docs/src/models/Heat1D.md
@@ -1,8 +1,8 @@
 # 1D Heat Equation
 
-Models are under an `Abstract_Models`
+Models are under an `Abstract_Model`
 ```@docs
-LiftAndLearn.Abstract_Models
+LiftAndLearn.Abstract_Model
 ```
 
 ```@autodocs
