+++ docs/src/refs.bib
@@ -106,3 +106,13 @@
   year={1993},
   publisher={Annual Reviews 4139 El Camino Way, PO Box 10139, Palo Alto, CA 94303-0139, USA},
 }
+
+
+@inproceedings{koike2024energy,
+  title={Energy-Preserving Reduced Operator Inference for Efficient Design and Control},
+  author={Koike, Tomoki and Qian, Elizabeth},
+  booktitle={AIAA SCITECH 2024 Forum},
+  pages={1012},
+  year={2024},
+  doi={https://doi.org/10.2514/6.2024-1012}
+}
