+++ src/ChaosGizmo/ChaosGizmo.jl
@@ -8,12 +8,9 @@ using LinearAlgebra
 using Parameters
 using SparseArrays
 
-# Import 
-#   - abstract type Abstract_Models 
-#   - operators
-# from LiftAndLearn
 import ..LiftAndLearn: Abstract_Model, operators, extractF, extractH
 
+include("integrators.jl")
 include("LyapunovExponent.jl")
 
 end
