+++ src/LiftAndLearn.jl
@@ -56,7 +56,7 @@ include("optimizer/NC_Optimize.jl")
 include("optimizer/EP_Optimize.jl")
 
 # Streaming
-include("Streaming/streaming.jl")
+# include("Streaming/streaming.jl")
 
 # [Submodule] Analysis of chaos analysis tools
 include("ChaosGizmo/ChaosGizmo.jl")
